task ,task , task, ... task ,task , task, ... task ,task , task, ... task ,task , task, ...

+++ trunk/Timesheet/src/be/model/Project.java
@@ -32,8 +32,8 @@ public class Project extends AbstractTask {
 	}
 
 	@Override
-	public Integer getWorkedHours() {
-		int result = 0;
+	public Double getWorkedHours() {
+		double result = 0;
 		for(AbstractTask child : this.getChildren()) {
 			result += child.getWorkedHours();
 		}
